requirements.txt
anthropic>=0.34.0
click>=8.1.7
rich>=13.7.0
python-dotenv>=1.0.0
Pillow>=10.2.0
streamlit>=1.28.0
pandas>=2.0.0

